+++ packages/core/src/charts/contexts.tsx
@@ -1,10 +1,7 @@
 import { createContext, useContext } from 'react'
-import { ChartDataContextProps, ChartDataProviderProps, ChartDataProviderValue } from './types'
+import { ChartDataProviderProps, ChartDataProviderValue } from './types'
 
-export const ChartDataContext = createContext({
-    data: {} as ChartDataContextProps,
-    setData: (data: ChartDataContextProps) => data,
-} as ChartDataProviderValue)
+export const ChartDataContext = createContext({ data: {} } as ChartDataProviderValue)
 
 export const ChartDataProvider = ({ value, children }: ChartDataProviderProps) => {
     return <ChartDataContext.Provider value={value}>{children}</ChartDataContext.Provider>